Update README.md
Association rule learning is a type of unsupervised learning technique that checks for the dependency of one data item on another data item and maps accordingly so that it can be more profitable. It tries to find some interesting relations or associations among the variables of dataset. It is based on different rules to discover the interesting relations between variables in the database.

The association rule learning is one of the very important concepts of machine learning, and it is employed in Market Basket analysis, Web usage mining, continuous production, etc. Here market basket analysis is a technique used by the various big retailer to discover the associations between items. We can understand it by taking an example of a supermarket, as in a supermarket, all products that are purchased together are put together.

Association rule learning works on the concept of If and Else Statement, such as if A then B. If element is called antecedent, and then statement is called as Consequent. These types of relationships where we can find out some association or relation between two items is known as single cardinality. It is all about creating rules, and if the number of items increases, then cardinality also increases accordingly. So, to measure the associations between thousands of data items, there are several metrics. These metrics are given below:

Support
Confidence
Lift
This assignment will study following Questions :
Statement No 1 : Prepare rules for the "books" data sets :

Statement No 2 : Prepare rules for the "my_movies" data sets :

Try different values of support and confidence. Observe the change in number of rules for different support, confidence values.

Change the minimum length in apriori algorithm.

Visulize the obtained rules using different plots.